feat(sales): extract useSales hook and SalesFormSection component
